+++ examples/swbd/s5c/exp/visualization/plot_hierarchical_ctc_probs.py
@@ -15,8 +15,6 @@ import shutil
 sys.path.append(abspath('../../../'))
 from models.load_model import load
 from examples.swbd.s5c.exp.dataset.load_dataset_hierarchical import Dataset
-from utils.io.labels.character import Idx2char
-from utils.io.labels.word import Idx2word
 from utils.directory import mkdir_join, mkdir
 from utils.visualization.ctc import plot_hierarchical_ctc_probs
 from utils.config import load_config
@@ -40,7 +38,7 @@ def main():
     params = load_config(join(args.model_path, 'config.yml'), is_eval=True)
 
     # Load dataset
-    test_data = Dataset(
+    dataset = Dataset(
         data_save_path=args.data_save_path,
         backend=params['backend'],
         input_freq=params['input_freq'],
@@ -54,8 +52,8 @@ def main():
         num_stack=params['num_stack'], num_skip=params['num_skip'],
         sort_utt=False, reverse=False, tool=params['tool'])
 
-    params['num_classes'] = test_data.num_classes
-    params['num_classes_sub'] = test_data.num_classes_sub
+    params['num_classes'] = dataset.num_classes
+    params['num_classes_sub'] = dataset.num_classes_sub
 
     # Load model
     model = load(model_type=params['model_type'],
@@ -68,67 +66,36 @@ def main():
     # GPU setting
     model.set_cuda(deterministic=False, benchmark=True)
 
-    # Visualize
-    plot(model=model,
-         dataset=test_data,
-         eval_batch_size=args.eval_batch_size,
-         save_path=mkdir_join(args.model_path, 'ctc_probs'))
+    save_path = mkdir_join(args.model_path, 'ctc_probs')
 
-
-def plot(model, dataset, eval_batch_size, save_path=None):
-    """
-    Args:
-        model: the model to evaluate
-        dataset: An instance of a `Dataset` class
-        eval_batch_size (int): the batch size when evaluating the model
-        save_path (string): path to save figures of CTC posteriors
-    """
-    # Set batch size in the evaluation
-    if eval_batch_size is not None:
-        dataset.batch_size = eval_batch_size
+    ######################################################################
 
     # Clean directory
-    if isdir(save_path):
+    if save_path is not None and isdir(save_path):
         shutil.rmtree(save_path)
         mkdir(save_path)
 
-    idx2word = Idx2word(dataset.vocab_file_path)
-    idx2char = Idx2char(dataset.vocab_file_path,
-                        capital_divide=dataset.label_type_sub == 'character_capital_divide')
-
     for batch, is_new_epoch in dataset:
-
         # Get CTC probs
         probs = model.posteriors(batch['xs'], batch['x_lens'],
                                  temperature=1)
-        probs_sub = model.posteriors(batch['xs'], batch['x_lens'], is_sub_task=True,
+        probs_sub = model.posteriors(batch['xs'], batch['x_lens'], task_index=1,
                                      temperature=1)
         # NOTE: probs: '[B, T, num_classes]'
         # NOTE: probs_sub: '[B, T, num_classes_sub]'
 
-        # Decode
-        best_hyps = model.decode(batch['xs'], batch['x_lens'],
-                                 beam_width=1)
-        best_hyps_sub = model.decode(batch['xs'], batch['x_lens'],
-                                     beam_width=1,
-                                     is_sub_task=True)
-
         # Visualize
         for b in range(len(batch['xs'])):
-
-            # Convert from list of index to string
-            str_hyp = idx2word(best_hyps[b])
-            str_hyp_sub = idx2char(best_hyps_sub[b])
-
             speaker = batch['input_names'][b].split('_')[0]
             plot_hierarchical_ctc_probs(
                 probs[b, :batch['x_lens'][b], :],
                 probs_sub[b, :batch['x_lens'][b], :],
                 frame_num=batch['x_lens'][b],
                 num_stack=dataset.num_stack,
-                str_hyp=str_hyp,
-                str_hyp_sub=str_hyp_sub,
-                save_path=mkdir_join(save_path, speaker, batch['input_names'][b] + '.png'))
+                spectrogram=batch['xs'][b, :, :dataset.input_freq],
+                save_path=mkdir_join(save_path, speaker,
+                                     batch['input_names'][b] + '.png'),
+                figsize=(40, 8))
 
         if is_new_epoch:
             break
